pyproject.toml
[tool.poetry]
name = "thumbnailer"
version = "0.1.0"
description = ""
authors = ["Edward Q. Bridges <github@eqbridges.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pillow = "^7.1.2"
python-magic = "^0.4.18"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
lgw = "^1.0.9"
boto3 = "^1.13.2"
pre-commit = "^2.3.0"

[tool.poetry.dev-dependencies.black]
version = "^19.3b0"
allow-prereleases = true
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
